pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "bookstore API"
authors = [
    {name = "Marlonn Locatelli",email = "marlonn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pytest-django (>=4.10.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
